requirements.txt
langgraph==1.0.1
langchain-core>=0.3.0
typing-extensions>=4.5.0
python-dotenv>=1.0.0
langchain-google-genai>=2.0.0
